Add archive page
Since the index page only lists the N most recent articles, there is now
way to easily lookup older articles.

This commit adds an archive page that lists all articles. The archive
page is linked at the bottom of the index page.

This commit also changes the number of articles shown on the index page
to 25 (was: 100). The RSS now contains all articles.

Fixes #34